fix(bridge): refinalize picks the same share card as the bridge (#804)
refinalize.ts fetched /api/activity/<garmin id>/image unconditionally,
the generic activity card, while the bridge picks the Hevy-enriched
workout card whenever workout_enrichment links the activity. Re-finalizing
a strength workout therefore swapped its gym card for a run-style card
with an empty route panel (the 2026-09-07 "Upper" push, re-finalized
after the photo miss that #798 addressed).

One shared selector (share-image.ts) now serves both entrypoints: hevy_id
lookup, three attempts with a 60 s timeout, and a note the caller logs.
The refinalize log names the card kind it fetched.

Closes #804
Closes #805
